+++ Assets/Scripts/Transaction.cs
@@ -7,16 +7,17 @@ using UnityEngine.Rendering.VirtualTexturing;
 
     public class Transaction
     {
+        //CHANGE QUANT TO INT 
         //Stock or crypto 
         public string transactionType;
         //buy or sell
         public string action;
         public string itemName;
-        public int quantity;
+        public decimal quantity;
         public decimal price;
         public DateTime transactionDate;
 
-        public Transaction(string transactionType,string action, string itemName, int quantity, decimal price, DateTime date)
+        public Transaction(string transactionType,string action, string itemName, decimal quantity, decimal price, DateTime date)
         {
             this.transactionType = transactionType;
             this.itemName = itemName;
